requirements.txt
# Core Office file parsing libraries
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=0.6.23

# Legacy Office format support and macro analysis
oletools>=0.60.1
olefile>=0.46

# Network and URL analysis
requests>=2.31.0
urllib3>=2.0.0
validators>=0.22.0

# Cryptographic and hashing
pycryptodome>=3.19.0

# Pattern matching and regex
regex>=2023.10.3

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0

# CLI and logging
click>=8.1.0
colorama>=0.4.6
tqdm>=4.66.0

# Type hints and development
typing-extensions>=4.8.0

# Testing and quality assurance
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0